reduce exports
