+++ jolt-core/src/subprotocols/sumcheck.rs
@@ -202,7 +202,6 @@ impl<F: PrimeField> SumcheckInstanceProof<F> {
     let mut cubic_polys: Vec<CompressedUniPoly<F>> = Vec::new();
 
     for _j in 0..params.num_rounds {
-
       #[cfg(feature = "multicore")]
       let iterator = params.pairs_par_iter();
 
@@ -211,7 +210,6 @@ impl<F: PrimeField> SumcheckInstanceProof<F> {
 
       let evals: Vec<(F, F, F)> = iterator
         .map(|(poly_A, poly_B, eq)| {
-
           let len = poly_A.len() / 2;
 
           #[cfg(feature = "multicore")]
@@ -220,40 +218,44 @@ impl<F: PrimeField> SumcheckInstanceProof<F> {
           #[cfg(not(feature = "multicore"))]
           let iterator = (0..len).into_par_iter();
 
-          let (eval_point_0, eval_point_2, eval_point_3) = iterator.map(|i| {
-            // eval 0: bound_func is A(low)
-            let eval_point_0 = params.combine(&poly_A[i], &poly_B[i], &eq[i]);
+          let (eval_point_0, eval_point_2, eval_point_3) = iterator
+            .map(|i| {
+              // eval 0: bound_func is A(low)
+              let eval_point_0 = params.combine(&poly_A[i], &poly_B[i], &eq[i]);
 
-            let high = len + i;
-            let low = i;
-            let m_a = poly_A[high] - poly_A[low];
-            let m_b = poly_B[high] - poly_B[low];
-            let m_eq = eq[high] - eq[low];
+              let high = len + i;
+              let low = i;
+              let m_a = poly_A[high] - poly_A[low];
+              let m_b = poly_B[high] - poly_B[low];
+              let m_eq = eq[high] - eq[low];
 
-            // eval 2: bound_func is -A(low) + 2*A(high)
-            let poly_A_bound_point = poly_A[high] + m_a;
-            let poly_B_bound_point = poly_B[high] + m_b;
-            let poly_C_bound_point = eq[high] + m_eq;
-            let eval_point_2 = params.combine(
-              &poly_A_bound_point,
-              &poly_B_bound_point,
-              &poly_C_bound_point,
+              // eval 2: bound_func is -A(low) + 2*A(high)
+              let poly_A_bound_point = poly_A[high] + m_a;
+              let poly_B_bound_point = poly_B[high] + m_b;
+              let poly_C_bound_point = eq[high] + m_eq;
+              let eval_point_2 = params.combine(
+                &poly_A_bound_point,
+                &poly_B_bound_point,
+                &poly_C_bound_point,
+              );
+
+              // eval 3: bound_func is -2A(low) + 3A(high); computed incrementally with bound_func applied to eval(2)
+              let poly_A_bound_point = poly_A_bound_point + m_a;
+              let poly_B_bound_point = poly_B_bound_point + m_b;
+              let poly_C_bound_point = poly_C_bound_point + m_eq;
+
+              let eval_point_3 = params.combine(
+                &poly_A_bound_point,
+                &poly_B_bound_point,
+                &poly_C_bound_point,
+              );
+
+              (eval_point_0, eval_point_2, eval_point_3)
+            })
+            .reduce(
+              || (F::zero(), F::zero(), F::zero()),
+              |(sum_0, sum_2, sum_3), (a, b, c)| (sum_0 + a, sum_2 + b, sum_3 + c),
             );
-
-            // eval 3: bound_func is -2A(low) + 3A(high); computed incrementally with bound_func applied to eval(2)
-            let poly_A_bound_point = poly_A_bound_point + m_a;
-            let poly_B_bound_point = poly_B_bound_point + m_b;
-            let poly_C_bound_point = poly_C_bound_point + m_eq;
-
-            let eval_point_3 = params.combine(
-              &poly_A_bound_point,
-              &poly_B_bound_point,
-              &poly_C_bound_point,
-            );
-
-            (eval_point_0, eval_point_2, eval_point_3)
-          })
-          .reduce(|| (F::zero(), F::zero(), F::zero()), |(sum_0, sum_2, sum_3), (a, b, c)| (sum_0 + a, sum_2 + b, sum_3 + c));
 
           (eval_point_0, eval_point_2, eval_point_3)
         })
@@ -603,6 +605,142 @@ impl<G: CurveGroup> ZKSumcheckInstanceProof<G> {
     }
 
     Ok((self.comm_evals[self.comm_evals.len() - 1], r))
+  }
+}
+
+pub mod bench {
+  use super::*;
+  use ark_curve25519::{EdwardsProjective, Fr};
+  use ark_std::{rand::Rng, test_rng, One, UniformRand, Zero};
+  use criterion::black_box;
+  use crate::poly::dense_mlpoly::DensePolynomial;
+  use crate::poly::eq_poly::EqPolynomial;
+  use crate::subprotocols::sumcheck::{CubicSumcheckParams, SumcheckInstanceProof};
+  use crate::utils::index_to_field_bitvector;
+
+  pub fn sumcheck_bench(
+    group: &mut criterion::BenchmarkGroup<'_, criterion::measurement::WallTime>,
+  ) {
+    let num_vars = 16;
+    let mut rng = test_rng();
+
+    // PLAIN
+    let r1 = vec![Fr::rand(&mut rng); num_vars];
+    let r2 = vec![Fr::rand(&mut rng); num_vars];
+    let r3 = vec![Fr::rand(&mut rng); num_vars];
+    let eq1 = DensePolynomial::new(EqPolynomial::new(r1).evals());
+    let eq2 = DensePolynomial::new(EqPolynomial::new(r2).evals());
+    let eq3 = DensePolynomial::new(EqPolynomial::new(r3).evals());
+    let params =
+      CubicSumcheckParams::new_prod(vec![eq1.clone()], vec![eq2.clone()], eq3.clone(), num_vars);
+
+    let mut claim = Fr::zero();
+    for i in 0..num_vars {
+      let eval1 = eq1.evaluate(&index_to_field_bitvector(i, num_vars));
+      let eval2 = eq2.evaluate(&index_to_field_bitvector(i, num_vars));
+      let eval3 = eq3.evaluate(&index_to_field_bitvector(i, num_vars));
+
+      claim += eval1 * eval2 * eval3;
+    }
+
+    let coeffs = vec![Fr::one()];
+
+    group.bench_function("sumcheck unbatched 2^16", |b| {
+      b.iter(|| {
+        let mut transcript = Transcript::new(b"test_transcript");
+        let params = black_box(params.clone());
+        let (proof, r, evals) = SumcheckInstanceProof::prove_cubic_batched_special::<
+          EdwardsProjective,
+        >(&claim, params, &coeffs, &mut transcript);
+      })
+    });
+
+    // FLAGGED
+    let num_leaves = 1 << num_vars;
+    let mut vals1 = vec![Fr::rand(&mut rng); num_leaves];
+    let mut vals2 = vec![Fr::rand(&mut rng); num_leaves];
+    // Set approximately half of the leaves to 1
+    for _ in 0..num_vars / 2 {
+      let rand_index = rng.gen_range(0..num_leaves);
+      vals1[rand_index] = Fr::one();
+      vals2[rand_index] = Fr::one();
+    }
+    let poly_a = DensePolynomial::new(vals1);
+    let poly_b = DensePolynomial::new(vals2);
+    let mut r = vec![Fr::rand(&mut rng); num_vars];
+    let eq = DensePolynomial::new(EqPolynomial::new(r).evals());
+    let params = CubicSumcheckParams::new_prod(
+      vec![poly_a.clone()],
+      vec![poly_b.clone()],
+      eq.clone(),
+      num_vars,
+    );
+
+    let mut claim = Fr::zero();
+    for i in 0..num_vars {
+      let eval1 = poly_a.evaluate(&index_to_field_bitvector(i, num_vars));
+      let eval2 = poly_b.evaluate(&index_to_field_bitvector(i, num_vars));
+      let eval3 = eq.evaluate(&index_to_field_bitvector(i, num_vars));
+
+      claim += eval1 * eval2 * eval3;
+    }
+
+    let coeffs = vec![Fr::one()];
+
+    group.bench_function("sumcheck unbatched (ones) 2^16", |b| {
+      b.iter(|| {
+        let mut transcript = Transcript::new(b"test_transcript");
+        let params = black_box(params.clone());
+        let (proof, r, evals) = SumcheckInstanceProof::prove_cubic_batched_special::<
+          EdwardsProjective,
+        >(&claim, params, &coeffs, &mut transcript);
+      })
+    });
+
+    // BATCHED
+    let batch_size = 10;
+    let num_vars = 14;
+
+    let r_eq = vec![Fr::rand(&mut rng); num_vars];
+    let eq = DensePolynomial::new(EqPolynomial::new(r_eq).evals());
+
+    let mut poly_as = Vec::with_capacity(batch_size);
+    let mut poly_bs = Vec::with_capacity(batch_size);
+    for _ in 0..batch_size {
+      let ra = vec![Fr::rand(&mut rng); num_vars];
+      let rb = vec![Fr::rand(&mut rng); num_vars];
+      let a = DensePolynomial::new(EqPolynomial::new(ra).evals());
+      let b = DensePolynomial::new(EqPolynomial::new(rb).evals());
+      poly_as.push(a);
+      poly_bs.push(b);
+    }
+    let params =
+      CubicSumcheckParams::new_prod(poly_as.clone(), poly_bs.clone(), eq.clone(), num_vars);
+    let coeffs = vec![Fr::rand(&mut rng); batch_size];
+
+    let mut joint_claim = Fr::zero();
+    for batch_i in 0..batch_size {
+      let mut claim = Fr::zero();
+      for var_i in 0..num_vars {
+
+        let eval_a = poly_as[batch_i].evaluate(&index_to_field_bitvector(var_i, num_vars));
+        let eval_b = poly_bs[batch_i].evaluate(&index_to_field_bitvector(var_i, num_vars));
+        let eval_eq = eq.evaluate(&index_to_field_bitvector(var_i, num_vars));
+
+        claim += eval_a * eval_b * eval_eq;
+      }
+      joint_claim += coeffs[batch_i] * claim;
+    }
+
+    group.bench_function("sumcheck 10xbatched 2^14", |b| {
+      b.iter(|| {
+        let mut transcript = Transcript::new(b"test_transcript");
+        let params = black_box(params.clone());
+        let (proof, r, evals) = SumcheckInstanceProof::prove_cubic_batched_special::<
+          EdwardsProjective,
+        >(&joint_claim, params, &coeffs, &mut transcript);
+      })
+    });
   }
 }
